# example-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для складов по городам (ДелаемIT)",
        "description": "Документация API для работы с городами и складами",
        "contact": {
            "email": "magoed694@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stocks": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Все склады",
                "description": "Получить список всех складов.",
                "operationId": "42b0e7ea63f4a60c43cc43767f758745",
                "responses": {
                    "200": {
                        "description": "Список складов.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Stock"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Создание склада",
                "description": "Создать новый склад.",
                "operationId": "deb5e4ef0bb32e7415b68fb9f456bc29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_id",
                                    "address",
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Плеханова, дом 2"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.323232
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.343432
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Склад успешно создан.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stock"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибки валидации."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stocks/{id}": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Просмотр склада",
                "description": "Получить информацию о складе по ID.",
                "operationId": "d5b06bf461a5b3eaecaaec3c277ff466",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID склада",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о складе.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stock"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Склад не найден."
                    }
                }
            },
            "put": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Обновление склада",
                "description": "Обновить информацию о складе.",
                "operationId": "b7d16c68813d3ba08518ad0f69eb5064",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID склада",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_id",
                                    "address",
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Плеханова, дом 2"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.323232
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 34.343432
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация о складе обновлена.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stock"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Склад не найден."
                    },
                    "422": {
                        "description": "Ошибки валидации."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Удаление склада",
                "description": "Удалить склад по ID.",
                "operationId": "0fca3aebabc6f2e43eff67a6c07ce940",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID склада",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Склад успешно удален."
                    },
                    "404": {
                        "description": "Склад не найден."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stocks/nearby": {
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Поиск ближайшего склада",
                "description": "Найти ближайший склад по координатам.",
                "operationId": "03e7151699b14255d1712ef8ab1164b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 23.323232
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 23.212143
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ближайший склад найден.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stock": {
                                            "$ref": "#/components/schemas/Stock"
                                        },
                                        "distance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5.6
                                        },
                                        "city": {
                                            "$ref": "#/components/schemas/City"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ближайший склад не найден."
                    },
                    "422": {
                        "description": "Ошибки валидации."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "City": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Minsk"
                    }
                },
                "type": "object"
            },
            "Stock": {
                "required": [
                    "id",
                    "city_id",
                    "name",
                    "lat",
                    "lng"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Warehouse"
                    },
                    "lat": {
                        "type": "numeric",
                        "example": "55,123456"
                    },
                    "lng": {
                        "type": "numeric",
                        "example": "23,123456"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Use Bearer Token for authorization",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}